refactored date parsing and checking to date_handler.rs
